requirements.txt
backcall==0.1.0
bleach==2.1.3
Brotli==1.0.4
decorator==4.3.0
dj-database-url==0.5.0
Django==2.0.7
django-cors-headers==2.4.0
django-filter==2.0.0
django-heroku==0.3.1
djangorestframework==3.8.2
entrypoints==0.2.3
graphviz==0.8.3
gunicorn==19.9.0
html5lib==1.0.1
ipykernel==4.8.2
ipython==6.4.0
ipython-genutils==0.2.0
ipywidgets==7.2.1
jedi==0.12.0
Jinja2==2.10
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==5.2.0
jupyter-core==4.4.0
Markdown==2.6.11
MarkupSafe==1.0
mistune==0.8.3
nbconvert==5.3.1
nbformat==4.4.0
notebook==5.5.0
numpy==1.14.5
opencv-contrib-python==3.4.1.15
pandocfilters==1.4.2
parso==0.2.1
pexpect==4.6.0
pickleshare==0.7.4
Pillow==5.1.0
prompt-toolkit==1.0.15
psycopg2==2.7.5
psycopg2-binary==2.7.5
ptyprocess==0.5.2
Pygments==2.2.0
pytesseract==0.2.2
python-dateutil==2.7.3
pytz==2018.5
pyzmq==17.0.0
qtconsole==4.3.1
raven==6.9.0
Send2Trash==1.5.0
simplegeneric==0.8.1
six==1.11.0
terminado==0.8.1
testpath==0.3.1
tornado==5.0.2
traitlets==4.3.2
wcwidth==0.1.7
webencodings==0.5.1
whitenoise==4.0
widgetsnbextension==3.2.1
